==> notebooks/RMarkdown/00_data_extraction.Rmd <==
---
title: "Data Extraction"
author: "Benzon Carlitos Salazar"
date: "`r Sys.Date()`"
output:
  html_document:
    code_folding: show
    df_print: paged
    toc: TRUE
    toc_float: TRUE
    theme: cosmo
---

```{r setup, include=FALSE}
library(magrittr)

full_data_folder <- paste0(
  "C:/Users/bsalaz5/Documents/Projects/School/kaggle-ASIA_Functional_Track/",
  "data-raw/ASIAChallenge_ShareFile"
)
```

## Data Extractions

Saving to local folder.

```{r, datasets}
metadata_df <-
  readr::read_csv(here::here(full_data_folder, "metadata.csv"),
                  show_col_types = FALSE)

train_features_df <-
  readr::read_csv(here::here(full_data_folder, "train_features.csv"),
                  show_col_types = FALSE)
train_outcomes_df <-
  readr::read_csv(here::here(full_data_folder, "train_outcomes_functional.csv"),
                  show_col_types = FALSE)

test_features_df <-
  readr::read_csv(here::here(full_data_folder, "test_features.csv"),
                  show_col_types = FALSE)
test_outcomes_df <-
  readr::read_csv(
    here::here(full_data_folder, "test_outcomes_Fun_template_update.csv"),
    show_col_types = FALSE
  )
```

```{r disconnect_save, include=FALSE}
metadata_df %>% nanoparquet::write_parquet(here::here("data-raw", "metadata.parquet"))
train_features_df %>% nanoparquet::write_parquet(here::here("data-raw", "train_features.parquet"))
train_outcomes_df %>% nanoparquet::write_parquet(here::here("data-raw", "train_outcomes.parquet"))
test_features_df %>% nanoparquet::write_parquet(here::here("data-raw", "test_features.parquet"))
test_outcomes_df %>% nanoparquet::write_parquet(here::here("data-raw", "test_outcomes.parquet"))

rm(list = ls(all.names = TRUE)) # Clear all objects including hidden objects
invisible(gc()) # free up memory
```
