==> notebooks/RMarkdown/01_eda.Rmd <==
---
title: "Exploratory Data Analysis"
author: "Benzon Carlitos Salazar"
date: "`r Sys.Date()`"
output:
  html_document:
    code_folding: show
    df_print: paged
    toc: TRUE
    toc_float: TRUE
    theme: cosmo
---

```{r setup, include=FALSE}
library(magrittr)
```

## EDA

### Metadata

- All metadata information are fixed information. These variables include:

| Variables    | Definition                                                                                  | Possible Values                     |
|--------------|---------------------------------------------------------------------------------------------|-------------------------------------|
| PID          | patient id                                                                                  |                                     |
| age_category | age at time of injury either above or below 65 years                                        | <65, >65                            |
| sexcd        | sex                                                                                         | 1, female; 2, male                  |
| bmi_category | bmi categorized                                                                             | underweight, healthy, overweight    |
| srdecc1      | were spinal roots decompressed during surgery                                               | 0, no; 1, yes                       |
| surgcd1      | was spinal surgery performed                                                                | 0, no; 1, yes                       |
| spcsuc1      | Are there more than one area of spinal column surgery                                       | 0, no; 1, yes                       |
| scdecc1      | was the spinal cord decompressed                                                            | 0, no; 1, yes                       |
| hemccd1      | did significant changes in the hemodynamic or pulmonary function occur during the operation | 0, no; 1, yes                       |
| mhpsyccd     | has the patient had a past history of psychiatric abnormalities                             | 0, no; 1, yes; 9, unknown           |
| mhneurcd     | has the patient had a past history of neurological abnormalities                            | 0, no; 1, yes; 9, unknown           |
| mhcardcd     | has the patient had a past history of cardiac abnormalities                                 | 0, no; 1, yes; 9, unknown           |
| mhmetacd     | has the patient had a past history of metabolic or endocrinologic abnormalities             | 0, no; 1, yes; 9, unknown           |
| tx1_r        | treatment group                                                                             | D1, dose 1; D2, dose 2; P, placebo  |


```{r metadata_summs}
metadata_df <- nanoparquet::read_parquet(here::here("data-raw", "metadata.parquet"))

metadata_proc <-
  metadata_df %>%
  dplyr::mutate(bmi_category = as.factor(ifelse(is.na(bmi_category), "Unknown", bmi_category)),
                age_category = as.factor(ifelse(is.na(age_category), "Unknown", age_category)),
                sexcd = as.factor(ifelse(is.na(sexcd), "Unknown", sexcd)),
                tx1_r = as.factor(ifelse(is.na(tx1_r), "Unknown", tx1_r)),
                srdecc1 = as.factor(ifelse(is.na(srdecc1), "Unknown", srdecc1)),
                surgcd1 = as.factor(ifelse(is.na(surgcd1), "Unknown", surgcd1)),
                spcsuc1 = as.factor(ifelse(is.na(spcsuc1), "Unknown", spcsuc1)),
                scdecc1 = as.factor(ifelse(is.na(scdecc1), "Unknown", scdecc1)),
                hemccd1 = as.factor(ifelse(is.na(hemccd1), "Unknown", hemccd1)),
                mhpsyccd = as.factor(ifelse(is.na(mhpsyccd), "Unknown", mhpsyccd)),
                mhneurcd = as.factor(ifelse(is.na(mhneurcd), "Unknown", mhneurcd)),
                mhcardcd = as.factor(ifelse(is.na(mhcardcd), "Unknown", mhcardcd)),
                mhmetacd = as.factor(ifelse(is.na(mhmetacd), "Unknown", mhmetacd)))

metadata_abt_row <-
  metadata_proc %>%
  dplyr::select(-PID) %>%
  gtsummary::tbl_summary(
    by = age_category,
    statistic = list(gtsummary::all_categorical() ~ "{n}/{N} ({p}%)"),
    percent = "row",
    sort = gtsummary::all_categorical() ~ "frequency",
    label = list(
      sexcd ~ "Sex",
      bmi_category ~ "BMI categorized ",
      srdecc1 ~ "were spinal roots decompressed during surgery",
      surgcd1 ~ "was spinal surgery performed",
      spcsuc1 ~ "Are there more than one area of spinal column surgery ",
      scdecc1 ~ "was the spinal cord decompressed",
      hemccd1 ~ "did significant changes in the hemodynamic or pulmonary function occur during the operation",
      mhpsyccd ~ "Has the patient had a past history of psychiatric abnormalities",
      mhneurcd ~ "Has the patient had a past history of neurological abnormalities",
      mhcardcd ~ "Has the patient had a past history of cardiac abnormalities",
      mhmetacd ~ "Has the patient had a past history of metabolic or endocrinologic abnormalities",
      tx1_r ~ "Treatment group"
    )
  ) %>%
  gtsummary::modify_header(label = "**Variable**",
                           stat_1 = "**<65 Years Old**<br>(N = {n})",
                           stat_2 = "**>65 Years Old**<br>(N = {n})") %>%
  gtsummary::bold_labels() %>%
  gtsummary::modify_footnote(update = dplyr::everything() ~ NA)

metadata_abt_col <-
  metadata_proc %>%
  dplyr::select(-c(PID, age_category)) %>%
  gtsummary::tbl_summary(
    statistic = list(gtsummary::all_categorical() ~ "{n}/{N} ({p}%)"),
    sort = gtsummary::all_categorical() ~ "frequency",
    label = list(
      sexcd ~ "Sex",
      bmi_category ~ "BMI categorized ",
      srdecc1 ~ "were spinal roots decompressed during surgery",
      surgcd1 ~ "was spinal surgery performed",
      spcsuc1 ~ "Are there more than one area of spinal column surgery ",
      scdecc1 ~ "was the spinal cord decompressed",
      hemccd1 ~ "did significant changes in the hemodynamic or pulmonary function occur during the operation",
      mhpsyccd ~ "Has the patient had a past history of psychiatric abnormalities",
      mhneurcd ~ "Has the patient had a past history of neurological abnormalities",
      mhcardcd ~ "Has the patient had a past history of cardiac abnormalities",
      mhmetacd ~ "Has the patient had a past history of metabolic or endocrinologic abnormalities",
      tx1_r ~ "Treatment group"
    )
  ) %>%
  gtsummary::modify_header(gtsummary::all_stat_cols() ~ "**Overall**<br>(N = {N})") %>%
  gtsummary::bold_labels() %>%
  gtsummary::modify_footnote(update = dplyr::everything() ~ NA)

metadata_abt__full <-
  gtsummary::tbl_merge(list(metadata_abt_col, metadata_abt_row)) %>%
  gtsummary::modify_caption("Metadata Information Across Known Age Groups At Time of Injury") %>%
  gtsummary::modify_spanning_header(dplyr::everything() ~ NA)

metadata_abt__full
```

### Training Features

```{r train_features}
train_features_df <- nanoparquet::read_parquet(here::here("data-raw", "train_features.parquet"))

# dimension
dim(train_features_df)
```

#### Columms with missing values

Top 10 missing values.

```{r train__missing_Values}
missing_values <- colSums(is.na(train_features_df))
missing_values_df <-
  data.frame(Column = names(missing_values),
             Missing_Values_Count = missing_values) %>%
  dplyr::arrange(desc(Missing_Values_Count))

missing_values_df %>% 
  dplyr::filter(Missing_Values_Count > 0) %>%
  head(n=10)
```

#### Low-variance features

```{r train__low_variance}
low_variance_cols <- 
  train_features_df %>%
  dplyr::select(dplyr::where(is.numeric)) %>%
  dplyr::summarise(dplyr::across(dplyr::everything(), sd, na.rm = TRUE)) %>%
  tidyr::pivot_longer(dplyr::everything(), names_to = "feature", values_to = "std_dev") %>%
  dplyr::filter(std_dev < 0.1)  # Threshold for low variance

low_variance_cols
```

- **No numeric columns have standard deviation below 0.01.**
- **No numeric columns have standard deviation below 0.05.**

Since the lowest standard deviation in the dataset is `0.06189822`, it means none of the numeric features have truly 
low variance under the typical near-zero variance threshold (e.g., < 0.01). This suggests that all features have some 
level of variability, and removing features based on variance thresholding alone may not be necessary.

**All features have some variability, meaning they could be useful for predictions.**

Possible concerns:

- High correlation?


#### Correlation analysis

```{r train__correlation_analysis}
corr_matrix <-
  cor(train_features_df %>%
        dplyr::select(dplyr::where(is.numeric)), use = "pairwise.complete.obs")
```

- Warning tells me that at least one column has a constant value. Correlation calculation fails for that column.

```{r train__correlation_analysis_cleanup}
constants <- 
  train_features_df %>%
  dplyr::select(dplyr::where(is.numeric)) %>%
  dplyr::summarise(dplyr::across(dplyr::everything(), sd, na.rm = TRUE)) %>%
  tidyr::pivot_longer(dplyr::everything(), names_to = "feature", values_to = "std_dev") %>%
  dplyr::filter(std_dev == 0)

constants

na_only_cols <- 
  train_features_df %>%
  dplyr::select(dplyr::where(is.numeric)) %>%
  dplyr::summarise(dplyr::across(dplyr::everything(), ~ all(is.na(.)))) %>%
  tidyr::pivot_longer(dplyr::everything(), names_to = "feature", values_to = "all_na") %>%
  dplyr::filter(all_na == TRUE)

na_only_cols

unique_counts <- 
  train_features_df %>%
  dplyr::select(dplyr::where(is.numeric)) %>%
  dplyr::summarise(dplyr::across(dplyr::everything(), ~ dplyr::n_distinct(na.omit(.)))) %>%
  tidyr::pivot_longer(dplyr::everything(), names_to = "feature", values_to = "unique_values") %>%
  dplyr::filter(unique_values == 1)

unique_counts

problematic_cols <- 
  train_features_df %>%
  dplyr::select(dplyr::where(is.numeric)) %>%
  dplyr::summarise(dplyr::across(dplyr::everything(), ~ sd(., na.rm = TRUE))) %>%
  tidyr::pivot_longer(dplyr::everything(), names_to = "feature", values_to = "std_dev") %>%
  dplyr::filter(is.na(std_dev) | std_dev == 0)

problematic_cols
```

No problematic columns...

```{r train__correlation_analysis_2, eval=FALSE}
corr_matrix <- 
  train_features_df %>%
  dplyr::select(dplyr::where(is.numeric)) %>%
  dplyr::select(dplyr::where(~ !all(is.na(.)))) %>%
  dplyr::select(dplyr::where(~ dplyr::n_distinct(na.omit(.)) > 1)) %>%
  cor(use = "pairwise.complete.obs")

high_corr <- caret::findCorrelation(corr_matrix, cutoff = 0.8)
df_reduced <- train_features_df[, -high_corr]  # Remove highly correlated features
```

Entirely possible that there are no highly correlated features.

```{r disconnect_save, include=FALSE}
rm(list = ls(all.names = TRUE)) # Clear all objects including hidden objects
invisible(gc()) # free up memory
```
