==> notebooks/RMarkdown/04_training_xgboost.Rmd <==
---
title: "XGBoost"
author: "Benzon Carlitos Salazar"
date: "`r Sys.Date()`"
output: 
  html_document:
    code_folding: hide
    df_print: paged
    toc: TRUE
    toc_float: TRUE
---

<style>
.main-container {
    width: 95%;
    max-width: unset;
}
</style>

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(magrittr)
training__PCA_df <- readRDS(here::here("data", "training__PCA.rds"))
```

```{r convertmodben}
training__PCA_df$modben <- as.factor(training__PCA_df$modben)
train_ready_df <- 
  training__PCA_df %>%
  dplyr::filter(!is.na(modben))
```

## Training using XGBoost

```{r training}
predictors <- train_ready_df %>% dplyr::select(PC1:PC5)
outcome <- train_ready_df$modben

# Split training (80%) and testing (20%)
set.seed(42)
train_index <- caret::createDataPartition(outcome, p = 0.8, list = FALSE)
train_set <- train_ready_df[train_index, ]
test_set <- train_ready_df[-train_index, ]
```

```{r xgboost}
set.seed(43)

# Convert labels to 0-based
train_labels <- as.numeric(train_set$modben) - 1
test_labels <- as.numeric(test_set$modben) - 1

train_matrix <- xgboost::xgb.DMatrix(data = as.matrix(train_set[, 1:5]), label = train_labels)
test_matrix <- xgboost::xgb.DMatrix(data = as.matrix(test_set[, 1:5]), label = test_labels)

# Model parameters
params <- list(
  objective = "multi:softmax",
  num_class = length(unique(train_set$modben)),
  eval_metric = "mlogloss",
  eta = 0.1,  # Learning rate
  max_depth = 6
)

xgb_model <- xgboost::xgb.train(params, train_matrix, nrounds = 100, watchlist = list(eval = test_matrix), print_every_n = 10L)

# Predict on test data
xgb_preds <- predict(xgb_model, test_matrix)

xgb_preds <- xgb_preds + 1  # Shift from 0-based to 1-based
caret::confusionMatrix(as.factor(xgb_preds), as.factor(test_set$modben))
```

### 1. Overall performance

- **Accuracy**: 0.5804 (58.04%)
- **Kappa score**: 0.3761
  - A Kappa between 0.2 and 0.4 suggests a fair level of agreement between predicted and actual labels but not strong.

```{r disconnect_save, include=FALSE}
rm(list = ls(all.names = TRUE)) # Clear all objects including hidden objects
invisible(gc()) # free up memory
```
