==> notebooks/RMarkdown/05_prediction__pca_rf.Rmd <==
---
title: "Tuned PCA RandomForest Training"
author: "Benzon Carlitos Salazar"
date: "`r Sys.Date()`"
output: 
  html_document:
    code_folding: hide
    df_print: paged
    toc: TRUE
    toc_float: TRUE
---

<style>
.main-container {
    width: 95%;
    max-width: unset;
}
</style>

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(magrittr)
rf_model <- readRDS(here::here("data/models", "RandomForest_tuned_model.rds"))

new_data <- 
  nanoparquet::read_parquet(here::here("data-raw", "test_outcomes.parquet")) %>%
  dplyr::select(-"time-DELETE THIS COLUMN FOR SUBMISSION") %>%
  dplyr::arrange(PID)
```

## New data predictions

```{r new_preds}
pids <- new_data$PID
new_data <- new_data %>% dplyr::select(-PID)
```

## PCA to new data

```{r pca_new}
new_pca_transformed <- as.data.frame(predict(rf_model, new_data))
```

```{r disconnect_save, include=FALSE}

rm(list = ls(all.names = TRUE)) # Clear all objects including hidden objects
invisible(gc()) # free up memory
```
